converted get purchases to server action
